o  fixes to Build queue.
git-svn-id: https://svn.apache.org/repos/asf/maven/continuum/branches/id-refactor@497932 13f79535-47bb-0310-9956-ffa450edef68

+++ continuum-core/src/main/java/org/apache/maven/continuum/buildqueue/evaluator/BuildProjectTaskViabilityEvaluator.java
@@ -75,7 +75,7 @@ public class BuildProjectTaskViabilityEvaluator
         {
             BuildProjectTask task = (BuildProjectTask) it.next();
 
-            Integer key = new Integer( task.getProjectId() );
+            Long key = new Long( task.getProjectId() );
 
             List projectTasks = (List) projects.get( key );
